devices/allwinner: Add a fel output
This is a self-contained (dependencies resolved through the nix store)
which allows booting the system through FEL.

This is not meant for distribution, but for testing.

I guess this could be changed to provide a more self-contained system,
but would probably require booting a static built sunxi-tools. For
another day!