wip/kernel: Workaround for inappropriate source re-use
The original eval's postInstall (with modules from Nixpkgs's config)
would be used and provided to the newly added postInstall step.

To properly fix this, I think we need to drop the Nixpkgs kernel builder
infra, and provide a submodule to build a kernel built onto a generic
modules-based mkDerivation submodule.